Update and rename p5.R to Banking project.R
A Portugese bank is rolling out term deposit for its customers. They have in the past connected to their customer base through phone calls. Results for these previous campaigns were recorded and have been provided to the current campaign manager to use the same in making this campaign more effective. 
Our goal is to  develop a statistical model which given this information predicts whether customer in question will subscribe to the product or not.